+++ pkg/util/googleutil/google.go
@@ -1,4 +1,4 @@
-package google
+package googleutil
 
 import (
 	"context"
